fix sla issues

+++ Data/Data/SASAml/SlaLevel3NonStaffViolatedWithoutAction.cs
@@ -7,8 +7,8 @@ namespace Data.Data.SASAml
     {
         public string? UserName { get; set; }
         public string? DisplayName { get; set; }
-        public long? AlertId { get; set; }
-        public long? CaseId { get; set; }
+        public decimal? AlertId { get; set; }
+        public decimal? CaseId { get; set; }
         public string? AlertedEntityNumber { get; set; }
         public string? AlertedEntityName { get; set; }
         public string? BranchName { get; set; }
