+++ Data/Data/SASAml/SlaLevel3StaffViolatedWithoutAction.cs
@@ -6,7 +6,7 @@ namespace Data.Data.SASAml
     public partial class SlaLevel3StaffViolatedWithoutAction
     {
         public string? UserName { get; set; }
-        public decimal? AlertId { get; set; }
+        public long? AlertId { get; set; }
         public DateTime? AssignedDate { get; set; }
         public string? AlertStatus { get; set; }
         public DateTime? AlertCloseDate { get; set; }
